+++ src/components/ProductPage/Hero.js
@@ -1,0 +1,37 @@
+import { StaticImage } from 'gatsby-plugin-image'
+import React from 'react'
+
+export default function Hero() {
+    return (
+        <section className="my-12 lg:my-0 grid grid-cols-1 lg:grid-cols-2 gap-7 lg:gap-14 items-center">
+            <div className="lg:max-w-[480px] justify-self-end px-5 box-content">
+                <h1 className="text-primary dark:text-white opacity-50 text-[20px] mb-0">What is PostHog?</h1>
+                <h2 className="text-[48px] mt-0 font-bold">
+                    An ever-expanding suite of tools to <span className="text-red">build better products</span>
+                </h2>
+                <ul className="text-[20px] font-semibold">
+                    <li className="text-[20px]">Measure performance</li>
+                    <li className="text-[20px]">Diagnose & build context</li>
+                    <li className="text-[20px]">Test & roll out improvements</li>
+                </ul>
+                <p>PostHog is open source, and a radically different approach to building better products.</p>
+                <p>
+                    Host on your own infrastructure and capture up to 30% more events, compared to cloud-based analytics
+                    tools.
+                </p>
+            </div>
+            <div className="justify-self-end relative flex flex-col">
+                <StaticImage loading="eager" quality={100} src="./images/hero-1.png" />
+                <div className="justify-self-end self-end w-[80%] mt-[-30%]">
+                    <StaticImage
+                        loading="eager"
+                        objectPosition="top"
+                        className="w-full"
+                        quality={100}
+                        src="./images/hero-2.png"
+                    />
+                </div>
+            </div>
+        </section>
+    )
+}
